pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "franklin-shifts"
version = "0.1.0"
description = "Analyze precinct-level partisan shifts in Franklin County, Ohio (2006→present)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Franklin Shifts Team"}
]
license = {text = "MIT"}

dependencies = [
    "geopandas>=0.14.0",
    "pandas>=2.1.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "rtree>=1.1.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "folium>=0.15.0",
    "contextily>=1.4.0",
    "mapclassify>=2.6.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ffs = "src.cli:app"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]


